% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatComp21019R.R
\name{Concentration}
\alias{Concentration}
\title{drug concentration}
\usage{
Concentration(
  t,
  d,
  t0 = c(1, 5, 9, 13, 17, 21, 25),
  Cl = 1.36 * 10,
  V = 3.4,
  tinf = rep(1e-05, 7)
)
}
\arguments{
\item{t}{the time}

\item{d}{the dose of drug}

\item{t0}{Administration time}

\item{Cl}{Clearance of elimination}

\item{V}{Volume of distribution}

\item{tinf}{the duration of the infusion of the jth administration}
}
\value{
the cytokine response
}
\description{
A 1-compartment infusion model for the PK
}
\examples{
\dontrun{
t<-seq(from=0,to=35,by=0.1)
plot(t,Concentration(t=t,d=rep(25,7)),type = "l",lty=2,ylab="Concentration ")
lines(t,Concentration(t=t,d=c(1,5,10,rep(25,4))),type = "l")
}
}
