% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatComp21019R.R
\name{Sa}
\alias{Sa}
\title{A dose regimen dataset}
\description{
A dataset used to show differnet  dose regimen.
}
\examples{
\dontrun{
data(Sa)
attach(Sa)
t<-seq(from=0,to=35,by=0.1)
plot(t,Concentration(t=t,d=rep(25,7)),type = "l",lty=2,ylab="Concentration ")
lines(t,Concentration(t=t,d=dose2),type = "l")
}
}
