\name{StatComp21019-package}
\alias{StatComp21019-package}
\docType{package}
\title{
Two functions for PK/PD models
}
\description{
The drug concentration function for the PK model and the cytokine response function for the PD model
}
\author{
Hao Liang
Maintainer: Hao Liang <loseriser@mail.ustc.edu.cn>
}
\keyword{ package }